requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Werkzeug==2.3.7
google-generativeai==0.3.2
Pillow==10.0.1
PyAutoGUI==0.9.54
PyMuPDF==1.23.8
python-docx==0.8.11
gunicorn==21.2.0 